# kanban-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kanban Board API",
        "description": "API para gestionar Tableros, Tarjetas y Tareas de un sistema Kanban, siguiendo la especificación OpenAPI 3.",
        "contact": {
            "name": "Tu Nombre",
            "email": "tu.email@ejemplo.com"
        },
        "license": {
            "name": "Licencia Estándar de API",
            "url": "http://localhost/licencia"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
        "/api/v1/boards": {
            "get": {
                "tags": [
                    "Tableros"
                ],
                "summary": "Obtiene todos los tableros",
                "operationId": "f7ecb9e5d3d3d6aa7f6081cd03d8e00d",
                "responses": {
                    "200": {
                        "description": "Lista de tableros obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Board"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tableros"
                ],
                "summary": "Crea un nuevo tablero",
                "operationId": "b768d6d0ee7d5a2a30bc2689bc6e61f7",
                "requestBody": {
                    "description": "Datos del tablero a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tablero de Marketing"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tablero creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Board"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/v1/boards/{id}": {
            "get": {
                "tags": [
                    "Tableros"
                ],
                "summary": "Muestra un tablero específico",
                "operationId": "9f4572331f161946dee42d3a7951f7dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tablero a mostrar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tablero encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Board"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tablero no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tableros"
                ],
                "summary": "Actualiza un tablero existente",
                "operationId": "981d46aeff4f10ce20e149c9d1574ff1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tablero a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del tablero para actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tablero de Marketing V2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tablero actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Board"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tablero no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tableros"
                ],
                "summary": "Elimina un tablero por su ID",
                "operationId": "b74ae434c2355f06084116a24d59ad90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tablero a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tablero eliminado exitosamente (No Content)"
                    },
                    "404": {
                        "description": "Tablero no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Board": {
                "title": "Board",
                "description": "Modelo de datos de un Tablero (Board)",
                "properties": {
                    "id": {
                        "description": "ID único del tablero",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del tablero (ej. Proyecto Alfa)",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}